backend-python/pyproject.toml
# SPDX-License-Identifier: MIT
# Copyright (c) 2025 Gabriel Xia(加百列)
[project]
name = "scaffold-backend-python"
version = "1.0.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.32.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "pydantic>=2.9.0",
]

